backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# LLM client (OpenAI-compatible, works with Qwen/DashScope)
openai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# HTTP client (for external APIs)
httpx>=0.27.0
